ai-service/requirements.txt
fastapi
uvicorn
pydantic-settings
python-jose